typo / optimisation / cleaning / removed warnings

+++ libraries/lib_pic/ccp.c
@@ -105,12 +105,12 @@ result_t ccp_set_pwm_duty (CCP_ID ccp_id, u16 duty)
     #if defined (_18F252)
 
         if (ccp_id == CCP_ID_1){
-            CCPR1L = (duty >> 2);                       // bits 9 to 2
+            CCPR1L = (u8)(duty >> 2);                   // bits 9 to 2
             CCP1CONbits.DC1B1 = (duty >> 1) & 0x01;     // bit 1
             CCP1CONbits.DC1B0 = duty & 0x01;            // bit 0
         }
         else if (ccp_id == CCP_ID_2){
-            CCPR2L = (duty >> 2);                       // bits 9 to 2
+            CCPR2L = (u8)(duty >> 2);                   // bits 9 to 2
             CCP2CONbits.DC2B1 = (duty >> 1) & 0x01;     // bit 1
             CCP2CONbits.DC2B0 = duty & 0x01;            // bit 0
         }
